# documentation/apidoc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the documentation for the GroupMaker project's backend API.",
    "version": "1.0.5",
    "title": "GroupMaker",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "ductran032118@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "groupmaker.swagger.io",
  "basePath": "/v1",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/accounts/<int:primary_key>/groups/": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "View all groups that include an user",
        "description": "",
        "operationId": "AccountGroupView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Showing accounts <id> groups",
              "id": "group.id",
              "name": "group.name"
            },
            "404": {
              "description": "EmptyGroupList: User <id> doesn't have any group"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: account",
              "read: group",
              "read: GroupMember"
            ]
          }
        ]
      }
    },
...
    "/groups/": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "View all created groups",
        "description": "",
        "operationId": "GroupView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [],
        "responses": [
          {
            "200": {
              "description": "Showing all groups available"
            },
            "404": {
              "description": "EmptyGroupList: No group created"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: account"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "group"
        ],
        "summary": "Register a new group",
        "description": "",
        "operationId": "GroupView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "group_name",
            "description": "Group's name",
            "required": true,
            "schema": {}
          }
        ],
        "responses": [
          {
            "201": {
              "description": "Group created"
            },
            "400": {
              "description": "Group name cannot be empty"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "write: group"
            ]
          }
        ]
      }
    },
    "/groups/<int:primary_key>/": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "View detailed information of a single group",
        "description": "",
        "operationId": "GroupDetailView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Showing group <id> details",
              "id": "group.id",
              "username": "group.name"
            },
            "404": {
              "description": "EmptyGrouprList: No group with this <id> found"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: group"
            ]
          }
        ]
      },
      "put": {
        "tags": [
          "group"
        ],
        "summary": "Modify details of a group",
        "description": "",
        "operationId": "GroupDetailView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          },
          {
            "in": "body",
            "name": "group_name",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Successfully modified group's information"
            },
            "404": {
              "description": "EmptyGroupList: No group found at <id>"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: group",
              "write: group"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "group"
        ],
        "summary": "Delete a group",
        "description": "",
        "operationId": "GroupDetailView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Successfully deleted group"
            },
            "404": {
              "description": "EmptyGroupList: No group found at <id>"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: group",
              "write: group"
            ]
          }
        ]
      }
    },
    "/groups/<int:primary_key>/members/": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "View the list of all members in group",
        "description": "",
        "operationId": "GroupMemberView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Showing group <id> member list",
              "id": "user.id",
              "username": "user.username",
              "email": "user.email",
              "token": "token.key"
            },
            "404": {
              "description": "EmptyUserList: This group does not have any user yet"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: GroupMember"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "group"
        ],
        "summary": "Add member to group",
        "description": "",
        "operationId": "GroupMemberView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          },
          {
            "in": "body",
            "name": "member_id",
            "required": true,
            "type": "int",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Successfully added member to group <id>"
            },
            "404": {
              "description": "EmptyUserList: No user found in group <id>"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: groupmember",
              "write: groupmember"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "group"
        ],
        "summary": "Remove a member from group",
        "description": "",
        "operationId": "AccountDetailView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Successfully deleted account"
            },
            "404": {
              "description": "EmptyUserList: No user found at <id>"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: group",
              "write: group"
            ]
          }
        ]
      }
    },
    "/groups/<int:primary_key>/tasks/": {
      "get": {
        "tags": [
          "group"
        ],
        "summary": "View the list of all tasks in group",
        "description": "",
        "operationId": "GroupTaskView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Showing group <id> tasks list",
              "id": "task.id",
              "task_name": "task.name",
              "task_description": "task.description",
              "task_created_time": "task.created_time",
              "task_deadline": "task.deadline"
            },
            "404": {
              "description": "EmptyTaskList: This group does not have any task yet"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: GroupTask"
            ]
          }
        ]
      },
      "post": {
        "tags": [
          "group"
        ],
        "summary": "Add task to group",
        "description": "",
        "operationId": "GroupTaskView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          },
          {
            "in": "body",
            "name": "task_id",
            "required": true,
            "type": "int",
            "minimum": 1,
            "description": ""
          },
          {
            "in": "body",
            "name": "task_name",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "in": "body",
            "name": "task_description",
            "required": false,
            "type": "string",
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Successfully added task to group <id>"
            },
            "404": {
              "description": "EmptyGroupList: No group found at <id>"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: grouptask",
              "write: grouptask"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "group"
        ],
        "summary": "Remove a task from group",
        "description": "",
        "operationId": "GroupTaskView",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "primary_key",
            "required": true,
            "type": "integer",
            "minimum": 1,
            "description": ""
          }
        ],
        "responses": [
          {
            "200": {
              "description": "Successfully deleted task"
            },
            "404": {
              "description": "EmptyTaskList: No task found at <id>"
            }
          }
        ],
        "security": [
          {
            "token_auth": [
              "read: grouptask",
              "write: grouptask"
            ]
          }
        ]
      }
    }
  },
...
}